requirements.txt
# Core dependencies
numpy>=1.23.0
scipy>=1.9.0

# Evolutionary algorithms
deap>=1.4.0

# Reinforcement learning environments
gymnasium>=0.29.0
mo-gymnasium>=1.0.0

# Neural network baselines
stable-baselines3>=2.0.0
torch>=2.0.0

# Machine learning utilities
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration management
pyyaml>=6.0

# Data management
pandas>=1.5.0

# Progress bars
tqdm>=4.64.0

# Testing
pytest>=7.2.0

# Optional: MuJoCo (install separately per official guide)
# mujoco>=2.3.0
